requirements.txt
numpy==1.25.2
pandas==2.0.3
Pillow==9.5.0
Pillow==10.2.0
pyttsx3==2.90
scikit_learn==1.3.0
streamlit==1.25.0
